+++ plugin.video.movie25/resources/libs/autoupdate.py
@@ -11,19 +11,16 @@ def unzipAndMove(_in, _out , src):
     except Exception, e:
         print str(e)
         return False
-
     return True
 
 def moveFiles(root_src_dir,root_dst_dir):
     for src_dir, dirs, files in os.walk(root_src_dir):
         dst_dir = src_dir.replace(root_src_dir, root_dst_dir)
-        if not os.path.exists(dst_dir):
-            os.mkdir(dst_dir)
+        if not os.path.exists(dst_dir): os.mkdir(dst_dir)
         for file_ in files:
             src_file = os.path.join(src_dir, file_)
             dst_file = os.path.join(dst_dir, file_)
-            if os.path.exists(dst_file):
-                os.remove(dst_file)
+            if os.path.exists(dst_file): os.remove(dst_file)
             shutil.move(src_file, dst_dir)
             
 def getUpdateFile(path,default = 0):
@@ -35,6 +32,5 @@ def getUpdateFile(path,default = 0):
     return default
 
 def saveUpdateFile(path,value):
-    try:
-        open(path,'w+').write(value)
+    try: open(path,'w+').write(value)
     except: pass
